requirement.txt
appdirs==1.4.4
asgiref==3.5.0
autopep8==1.5.5
distlib==0.3.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
filelock==3.0.12
importlib-metadata==4.11.3
Markdown==3.3.6
pbr==5.5.1
pycodestyle==2.6.0
Pygments==2.12.0
pytz==2022.1
six==1.15.0
sqlparse==0.4.2
stevedore==3.3.0
toml==0.10.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
zipp==3.8.0
